+++ docs/source/api_ref_utilities.rst
@@ -133,4 +133,3 @@ Miscellaneous
     set_seed
     generate
     torch_version_ge
-    TuneRecipeArgumentParser
